API 매니저 수정
